src/requirements.txt
pandas~=1.0.3
mysql-connector-python~=8.0.18
python-dateutil~=2.8.1
pydriller~=1.14
requests~=2.23.0
ujson~=2.0.3
spacy~=2.2.3
pygithub~=1.47